Add 15 unit tests
we add 15 unit tests in our project ( the AllTests. java)
test the project is training complete or not
test the layer number
test the size
test the predict List size
test the 0-9 elements in the training List is correct or not.  And an extra test for 10 ( obviously, this one did not pass, in our project, we don't have 10 )